Add files via upload
Delta Hedge Paper Trade: 

Bought at-the-money SPY call contracts with expiration December 16, 2022.  Choose the strike to be the strike closest to the current share price of SPY.  Bought as many contracts as I could, such that the total price ("premium") that you pay for the contracts is less than $10,000. Sold shares of SPY to delta-hedge my option contracts, meaning that the total delta of my SPY call options and ,y SPY stock should be as close as possible to zero. Used the delta calculated by IB TWS.