Add files via upload
Bought 800 shares of AAPL stock. . 
Bought AAPL put options to ensure that the combined value of your AAPL stock + AAPL puts
on January 20, 2023 will not be less than 90% of the current value of my AAPL stock.
Subject to that constraint, chose the strike of the AAPL put and the number of AAPL put contracts,
to minimize the amount I paid for this put "protection".

Sold AAPL call contracts with the same expiry as the puts,
and the same number of contracts as the puts, with call strike chosen
so that the price (the "premium") that you collect for the call
options is more than what you paid for the put options.
Subject to that constraint, I chose the call strike as high as possible.

This limits my downside risk to 10%, but it also caps my upside.